requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
anthropic==0.40.0
twilio==9.3.5
sqlalchemy==2.0.36
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
reportlab==4.2.5
python-multipart==0.0.12
httpx==0.27.2
